Complete insertion tests

+++ src/utils/hashing.rs
@@ -1,6 +1,6 @@
 use std::hash::{DefaultHasher, Hash, Hasher};
-pub fn calculate_hash<T: Hash>(t: &T) -> u8 {
+pub fn calculate_hash<T: Hash>(t: &T) -> usize {
     let mut hasher = DefaultHasher::new(); // Create a new DefaultHasher
     t.hash(&mut hasher); // Hash the value
-    hasher.finish() as u8 // Get the resulting hash
+    hasher.finish() as usize // Get the resulting hash
 }
